Refactor files
Belle.java is not assigned in a package.

Place Belle.java in package duke.

It is a requirement for classes to be in a package.
