+++ src/main/data/belleList.txt
@@ -1,14 +1,12 @@
-D , false , return book , 2/12/2019 1800
 T , false , abc
-T , false , happy
 E , false , project meeting , Mon 2pm , 4pm
-D , true , return book , Oct 15 2019
 T , false , abc
 T , false , abc
 E , false , project meeting , Mon 2pm , 4pm
 T , false , abc
 E , true , project meeting , Mon 2pm , 4pm
 T , false , abc
-D , false , return book , Sunday
-D , false , swim , 6pm
-E , true , project meeting , Mon 2pm , 4pm
+D , true , swim , 6pm
+E , false , project meeting , Mon 2pm , 4pm
+T , false , eat ban
+E , true , pm event project meeting , Mon 2pm , 4pm
