+++ src/main/java/duke/others/BelleException.java
@@ -1,4 +1,4 @@
-package others;
+package duke.others;
 public class BelleException extends Exception {
     public BelleException(String msg) {
         super(msg);
