+++ text-ui-test/EXPECTED.TXT
@@ -2,3 +2,6 @@
 Welcome!! I'm Belle <3.
 What can I do for you?
 --------------------------
+--------------------------
+Till next time!! Goodbye.
+--------------------------
