+++ text-ui-test/input.txt
@@ -1,0 +1,1 @@
+bye
